Moved some classes to handlersAndAction-package. Also moved Mahmoud's consumables bug fix.
