+++ src/view/GameOverScreen.java
@@ -1,7 +1,7 @@
 package view;
 
 import controller.GameLogic;
-import controller.handlers.ActionHandler;
+import controller.handlersAndActions.ActionHandler;
 import javax.swing.*;
 import java.awt.*;
