+++ src/view/creators/ObjectCreator.java
@@ -3,7 +3,7 @@ package view.creators;
 import controller.GameLogic;
 import controller.ImageResizer;
 import view.MainFrame;
-import controller.handlers.ActionHandler;
+import controller.handlersAndActions.ActionHandler;
 
 import javax.swing.*;
 import java.awt.event.MouseEvent;
