+++ src/view/creators/SceneCreator.java
@@ -2,7 +2,7 @@ package view.creators;
 
 import controller.ImageResizer;
 import view.MainFrame;
-import controller.handlers.ActionHandler;
+import controller.handlersAndActions.ActionHandler;
 import javax.swing.*;
 import java.io.BufferedReader;
 import java.io.FileInputStream;
