+++ src/view/panels/ProfilesPanel.java
@@ -1,6 +1,6 @@
 package view.panels;
 
-import controller.handlers.ActionHandler;
+import controller.handlersAndActions.ActionHandler;
 
 import javax.swing.*;
 import java.awt.*;
